Feature: refactor NewsDetail component and related files for improved structure and readability
